fix: create shared sanitizeForLog utility to prevent recurring CodeQL alerts
This addresses the root cause of recurring CodeQL log-injection alerts:

1. Created scripts/sanitize-for-log.js as a shared utility module
2. Updated scripts/update-security-status.js and scripts/test-pages-http.js to use it
3. This ensures CodeQL can trace sanitization through a recognized module

Why this keeps happening:
- CodeQL flags ANY user input (error.message, url, etc.) in console calls
- Even with sanitization, CodeQL needs to recognize the sanitization pattern
- Using a shared module helps CodeQL trace the sanitization flow

This should prevent future alerts #61, #62 and similar issues.

Fixes alerts #61 and #62
